update examples move iter package to internal
